chore: bump typescript + fix typings + fix circular deps
