+++ apps/balances-demo/package.json
@@ -38,7 +38,7 @@
     "eslint": "^7.5.0",
     "postcss": "^8.4.33",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.2.2",
+    "typescript": "^5.6.3",
     "vite": "^4.4.3"
   },
   "eslintConfig": {
