+++ packages/balances/package.json
@@ -43,7 +43,7 @@
     "polkadot-api": "^1.0.1",
     "rxjs": "^7.8.1",
     "scale-ts": "^1.6.0",
-    "viem": "^2.21.28"
+    "viem": "^2.21.32"
   },
   "devDependencies": {
     "@polkadot/api-contract": "14.1.1",
@@ -59,7 +59,7 @@
     "eslint": "^8.52.0",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "peerDependencies": {
     "@polkadot/api-contract": "12.0.2",
