+++ packages/chain-connector-evm/package.json
@@ -30,7 +30,7 @@
     "@talismn/util": "workspace:*",
     "anylogger": "^1.0.11",
     "lodash": "4.17.21",
-    "viem": "^2.21.28"
+    "viem": "^2.21.32"
   },
   "devDependencies": {
     "@talismn/eslint-config": "workspace:*",
@@ -40,7 +40,7 @@
     "eslint": "^8.52.0",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "eslintConfig": {
     "root": true,
