+++ packages/chaindata-provider/package.json
@@ -42,7 +42,7 @@
     "prettier": "^2.8.8",
     "ts-jest": "^29.1.1",
     "ts-node": "^10.9.2",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "preconstruct": {
     "entrypoints": [
