+++ packages/extension-core/package.json
@@ -19,10 +19,11 @@
     "clean": "rm -rf dist background/dist .turbo node_modules"
   },
   "dependencies": {
+    "@mangata-finance/type-definitions": "2.1.2",
     "@metamask/browser-passworder": "4.3.0",
     "@metamask/eth-sig-util": "7.0.1",
     "@polkadot-api/merkleize-metadata": "^1.1.7",
-    "@polkadot/apps-config": "0.132.1",
+    "@polkadot/apps-config": "0.145.1",
     "@polkadot/extension-base": "0.54.1",
     "@polkadot/extension-inject": "0.54.1",
     "@polkadot/keyring": "13.2.1",
@@ -64,7 +65,7 @@
     "ts-results": "3.3.0",
     "url-join": "^5.0.0",
     "uuid": "^8.3.2",
-    "viem": "^2.21.28",
+    "viem": "^2.21.32",
     "yup": "0.32.11"
   },
   "devDependencies": {
@@ -87,7 +88,7 @@
     "jest-webextension-mock": "3.8.9",
     "sinon-chrome": "^3.0.1",
     "ts-jest": "29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "preconstruct": {
     "entrypoints": [
