+++ packages/icons/package.json
@@ -37,7 +37,7 @@
     "jest": "^29.7.0",
     "react": "18.2.0",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "peerDependencies": {
     "react": ">=18.2.0"
