+++ packages/orb/package.json
@@ -33,7 +33,7 @@
     "eslint": "^8.52.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "peerDependencies": {
     "react": "*",
