+++ packages/scale/package.json
@@ -39,7 +39,7 @@
     "eslint": "^8.52.0",
     "jest": "^29.7",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "eslintConfig": {
     "root": true,
