+++ packages/talisman-ui/package.json
@@ -40,7 +40,7 @@
     "react-router-dom": "6.14.1",
     "react-use": "17.4.0",
     "tailwindcss": "^3.4.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "peerDependencies": {
     "@floating-ui/react": ">=0.20.1",
