+++ packages/token-rates/package.json
@@ -36,7 +36,7 @@
     "eslint": "^8.52.0",
     "jest": "^29.7.0",
     "ts-jest": "^29.1.1",
-    "typescript": "^5.2.2"
+    "typescript": "^5.6.3"
   },
   "eslintConfig": {
     "root": true,
