+++ packages/util/src/twox64Concat.ts
@@ -4,5 +4,6 @@ import { xxhashAsU8a } from "@polkadot/util-crypto"
 const bitLength = 64
 
 export function twox64Concat(input: string | Buffer | Uint8Array): `0x${string}` {
-  return u8aToHex(u8aConcat(xxhashAsU8a(input, bitLength), u8aToU8a(input)))
+  const inputAsU8a = typeof input === "string" ? input : new Uint8Array(input)
+  return u8aToHex(u8aConcat(xxhashAsU8a(inputAsU8a, bitLength), u8aToU8a(inputAsU8a)))
 }
